fixed login (I think), moved db helpers to own module
